use axios instead of fetch API
- Use consitional rendering to replace views based on global stage

Error: Objects are not valid as a React child (found: object with keys {name, age, city}).
If you meant to render a collection of children, use an array instead.

Objects are not valid as a React child (found: object with keys {users}). If you meant to render a collection of children, use an array instead.